Uniform hvl_tables.c to one common source
