+++ hvl2wav/makefile
@@ -41,7 +41,7 @@ hvl2wav.o: hvl2wav.c hvl_replay.h ../hvl_replay.h types.h
 hvl_replay.o: hvl_replay.c hvl_replay.h ../hvl_replay.h hvl_tables.h ../hvl_tables.h types.h
 	$(CC) -c hvl_replay.c -o hvl_replay.o $(CFLAGS)
 
-hvl_tables.o: hvl_tables.c hvl_tables.h ../hvl_tables.h types.h
+hvl_tables.o: hvl_tables.c ../hvl_tables.c hvl_tables.h ../hvl_tables.h types.h
 	$(CC) -c hvl_tables.c -o hvl_tables.o $(CFLAGS)
 
 clean:
